.requirements.txt
beautifulsoup4
mcstatus
tweepy
wolframalpha
python-dateutil
pyfiglet
bottle
dnspython
ipython
